+++ CHANGELOG.md
@@ -8,7 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
 - Add the `Gas` class
+- Add the `getNetworkBlockHeight` method for the `Wallet` class
+- Add tests for the new `getNetworkBlockHeight` method
 
 ### Removed
 
@@ -17,27 +20,33 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### [0.4.2]
 
 ### Changed
+
 - Change MAX_KEYS to 3 in wallet-core
 - Update version of wallet-core to 0.21.9
 
 ### Added
+
 - Add signal to sync to make it abortable
 
 ### [0.4.1]
 
 ### Removed
+
 - stake-allow tx removed
 
 ### Added
+
 - Integrate `exu` library for supporting web-workers
 - Add `WALLET_CORE_PATH` env variable for building
 - Add tests for npm.js
 
 ### Changed
+
 - Update CI to build and link with wallet-core on actions with the `WALLET_CORE_PATH` env variable
 
 ### [0.3.2]
 
 ### Changed
+
 - Ship the web assembly embedded in the library itself
 - During sync only, remove json from web assembly calls and send raw buffers